# server/openapi_types.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package server

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// GetSession defines model for GetSession.
type GetSession struct {
	Values *map[string]interface{} `json:"values,omitempty"`
}

// PostSession defines model for PostSession.
type PostSession struct {
	SessionKey   string                 `json:"sessionKey"`
	SessionValue map[string]interface{} `json:"sessionValue"`
}

// SetSessionJSONBody defines parameters for SetSession.
type SetSessionJSONBody = PostSession

// SetSessionJSONRequestBody defines body for SetSession for application/json ContentType.
type SetSessionJSONRequestBody = SetSessionJSONBody